pyproject.toml
[project]
name = "ai-agent-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
